fix: Scroll when mouse is not over the content
The version 3.0.1 fixed the navbar glitch, but introduced another bug: when the mouse was not over the page content it was not possible to scroll.

This patch corrects that partially. Meaning, now the page can be scrolled again, even if the mouse is not over the content. However, if the mouse is over the navbar, the scroll will not work. Because `base.css` sets the `pointer-events` to `none` in `stAppViewContainer`.

A future update is needed to address this. In order to keep the menu button clickable, and still be able to scroll the page, even when the pointer is over the navbar.