fix: Correct iframe height bug
On rare occasions, when the page is loaded, the function `Streamlit.setFrameHeight()` could not find the height of the widget and used the default value of 150px instead. Which lead to the iframe window being above the body of the web page.

This patch adjusts the iframe height in the `base.css`, to make sure that this will not happen.