feat: Remove max version requirement
The Streamlit team launched version 1.33.0, which works with `st_navbar` again. So there was probably a bug in version 1.32.2 that made the navbar not render properly. This commit removes the max version dependency, and instead excludes just the specific 1.32.2 version. Opening the package to the most up to date Streamlit features.